requirements.txt
absl-py==2.1.0
altair==4.0.0
astor==0.8.1
astunparse==1.6.3
attrs==23.2.0
base58==2.1.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
entrypoints==0.4
flatbuffers==23.5.26
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.3
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==3.19.6
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
